Ajustes en el registro y login - incremento 4

+++ database/migrations/2020_09_10_022906_create_ingresos_table.php
@@ -20,7 +20,7 @@ class CreateIngresosTable extends Migration
             $table->string('servicio');
             $table->string('detalle')->nullable();
             $table->string('tipo');
-            $table->integer('precio');
+            $table->bigInteger('precio');
             $table->timestamps();
             $table->softDeletes();
         });
